+++ ext.php
@@ -12,13 +12,13 @@ namespace primehalo\primenotify;
 
 class ext extends \phpbb\extension\base
 {
-	private static $conversion_table = array(
-				'primehalo.primenotify.notification.type.pm'				=> 'notification.type.pm',
-				'primehalo.primenotify.notification.type.post'				=> 'notification.type.post',
-				'primehalo.primenotify.notification.type.topic'				=> 'notification.type.topic',
+	private static $notification_types = array(
+				'primehalo.primenotify.notification.type.pm'	=> 'notification.type.pm',
+				'primehalo.primenotify.notification.type.post'	=> 'notification.type.post',
+				'primehalo.primenotify.notification.type.topic'	=> 'notification.type.topic',
 			);
 	private $db = null;
-	private $phpbb_notifications = null;
+	//private $phpbb_notifications = null;
 
     public function is_enableable()
     {
@@ -27,8 +27,11 @@ class ext extends \phpbb\extension\base
     }
 
 	/**
-	* Enable our notifications, disable the original notifications that we are
-	* replacing, and change user notification types to ours.
+	* When enabling this extension we want to make a primenotify copy of each
+	* equivalent notification type so that it will be enabled or disabled
+	* according to the user's current notification preferences.
+	* If we did not do this, the user's notification settings would not be
+	* enabled until they specifically enabled them from the UCP.
 	*
 	* @param mixed $old_state State returned by previous call of this method
 	* @return mixed Returns false after last step, otherwise temporary state
@@ -38,9 +41,35 @@ class ext extends \phpbb\extension\base
 	{
         if ($old_state === false)
         {
-			foreach (self::$conversion_table as $new => $old)
+			$this->db = !$this->db ? $this->container->get('dbal.conn') : $this->db;
+
+			// Grab all of our existing custom notifications (there shouldn't be any but we have to make sure otherwise we'll get an SQL error in the next step)
+			$sql = 'SELECT * FROM ' . USER_NOTIFICATIONS_TABLE . ' WHERE item_type ' . $this->db->sql_like_expression('primehalo.primenotify.notification.type.' . $this->db->get_any_char());
+			$result = $this->db->sql_query($sql);
+			while ($row = $this->db->sql_fetchrow($result))
 			{
-				$this->update_notification_type($new, $old);
+				$user_ids_per_type[$row['item_type']][] = $row['user_id'];
+			}
+
+			// Now loop through each notification type that has an equivalent primenotify type so that we can make a primenotify copy of it
+			foreach (self::$notification_types as $our_type => $orig_type)
+			{
+				// Create our notification type entries for each equivalent default notification type that already exists, copying over the notify status setting
+				$sql = 'SELECT * FROM ' . USER_NOTIFICATIONS_TABLE . " WHERE item_type = '{$orig_type}' AND method = 'notification.method.email'";
+				$result = $this->db->sql_query($sql);
+				if ($result)
+				{
+					$sql_ary = array();
+					while ($row = $this->db->sql_fetchrow($result))
+					{
+						if (!isset($user_ids_per_type[$our_type][$row['user_id']]))
+						{
+							$sql_ary[] = array('item_type' => $our_type, 'user_id' => $row['user_id'], 'method' => 'notification.method.email', 'notify' => $row['notify']);
+						}
+					}
+					$this->db->sql_multi_insert(USER_NOTIFICATIONS_TABLE, $sql_ary);
+				}
+
 			}
 
 			// Purge the cache to make sure our notification types show up on the UCP instead of the original types
@@ -54,8 +83,8 @@ class ext extends \phpbb\extension\base
 	}
 
 	/**
-	 * Disable our notifications, enable the original notifications that we
-	 * replaced, and change user notification types back to the originals.
+	 * When disabling this extension we need to convert existing primenotify notifications
+	 * to their phpBB equivalents and delete all primenotify notification types.
 	 *
 	 * @param mixed $old_state State returned by previous call of this method
 	 * @return mixed Returns false after last step, otherwise temporary state
@@ -65,11 +94,45 @@ class ext extends \phpbb\extension\base
 	{
         if ($old_state === false)
         {
-			foreach (self::$conversion_table as $new => $old)
+			$this->db = !$this->db ? $this->container->get('dbal.conn') : $this->db;
+
+			// Convert all existing custom notifications into default notifications
+			// Part 1/4: Get the notification type IDs so we know which notification IDs to find and what to convert them to
+			$sql_in = array_merge(array_values(self::$notification_types), array_keys(self::$notification_types));
+			$sql = 'SELECT * FROM ' . NOTIFICATION_TYPES_TABLE . ' WHERE ' . $this->db->sql_in_set('notification_type_name', $sql_in);
+			$result = $this->db->sql_query($sql);
+
+			// Part 2/4: Organize the notification IDs for easy access, using the type_name as the key and the type_id as the value
+			$notification_type_ids = array();
+			while ($row = $this->db->sql_fetchrow($result))
 			{
-				$this->update_notification_type($old, $new);
+				$notification_type_ids[$row['notification_type_name']] = $row['notification_type_id'];
 			}
 
+			// Part 3/4: Change notification_type_id from our custom one to the equivalent default one
+			foreach (self::$notification_types as $from => $to)
+			{
+				if (isset($notification_type_ids[$to]) && isset($notification_type_ids[$from]))
+				{
+					$sql = 'UPDATE ' . NOTIFICATIONS_TABLE . " SET notification_type_id = {$notification_type_ids[$to]} WHERE {$notification_type_ids[$from]}";
+					$this->db->sql_query($sql);
+				}
+			}
+
+			// Part 4/4: All notifications should have been converted, but just in case lets try deleting all notifications still containing our custom notification IDs
+			$sql = 'DELETE t1 FROM ' . NOTIFICATIONS_TABLE . ' t1 JOIN ' . NOTIFICATION_TYPES_TABLE . ' t2 ON t1.notification_type_id = t2.notification_type_id AND t2.notification_type_name ' . $this->db->sql_like_expression('primehalo.primenotify.notification.type.' . $this->db->get_any_char());
+			$this->db->sql_query($sql);
+
+			// Delete our custom notification types from the Notification Types Table
+			$sql = 'DELETE FROM ' . NOTIFICATION_TYPES_TABLE . ' WHERE notification_type_name ' . $this->db->sql_like_expression('primehalo.primenotify.notification.type.' . $this->db->get_any_char());
+			$this->db->sql_query($sql);
+
+			// Delete our custom notification types from the User Notifications Table
+			$sql = 'DELETE FROM ' . USER_NOTIFICATIONS_TABLE . ' WHERE item_type ' . $this->db->sql_like_expression('primehalo.primenotify.notification.type.' . $this->db->get_any_char());
+			$result = $this->db->sql_query($sql);
+			//file_put_contents("D:/output.txt", "sql=$sql, result=$result\n", FILE_APPEND);
+
+			// Clear the cache
 			$cache = $this->container->get('cache');
 			$cache->purge();
 
@@ -77,37 +140,5 @@ class ext extends \phpbb\extension\base
         }
 
         return parent::disable_step($old_state);
-	}
-
-	/**
-	* Enables and disables corresponding notification types.
-	*
-	* @param string $enable		Name of the notification to enable
-	* @param string $disable	Name of the notification to disable
-	* @return nothing
-	* @access private
-	*/
-	private function update_notification_type($enable, $disable)
-	{
-		if (!$this->phpbb_notifications)
-		{
-			$this->phpbb_notifications = $this->container->get('notification_manager');
-		}
-		if (!$this->db)
-		{
-			$this->db = $this->container->get('dbal.conn');
-		}
-
-		// Enable our notifications
-		$this->phpbb_notifications->enable_notifications($enable);
-
-		// Disable original notifications
-		$this->phpbb_notifications->disable_notifications($disable);
-
-		// Change user's original notifications to ours
-		$data = array('item_type' => $enable);
-		$sql = 'UPDATE ' . USER_NOTIFICATIONS_TABLE . ' SET ' . $this->db->sql_build_array('UPDATE', $data) .
-				 " WHERE item_type = '{$disable}' AND method = 'notification.method.email'";
-		$this->db->sql_query($sql);
 	}
 }